1. Introduce DB schema upgrade 2. Update to TWS 1020.01 version 3. Config in database support (ConfigRepo) 4. Historical data endpoint /historical 5. forceful run of scheduled task endpoint (/runTask)

+++ app/src/main/scala/serv1/db/schema/TickerDataTableGen.scala
@@ -32,7 +32,7 @@ class TickerDataTableGen(tt: TickerLoadType) {
 
     def close = column[Long]("CLOSE")
 
-    def volume = column[Long]("VOL")
+    def volume = column[Double]("VOL")
 
     def * = (id, time, open, high, low, close, volume) <> ((TickerData.apply _).tupled, TickerData.unapply)
